Conflicts:
	Projects/TestBed/Classes/Infrastructure/TeamCityTestsOutput.cpp
	Projects/TestBed/Classes/Infrastructure/UITestTemplate.h
	Projects/TestBed/Classes/OldTests/AlignTest.cpp
	Projects/TestBed/Classes/OldTests/AlignTest.h
	Projects/TestBed/Classes/OldTests/BiDiTest.cpp
	Projects/TestBed/Classes/OldTests/BiDiTest.h
	Projects/TestBed/Classes/OldTests/CloneTest.cpp
	Projects/TestBed/Classes/OldTests/CloneTest.h
	Projects/TestBed/Classes/OldTests/DLCDownloadTest.cpp
	Projects/TestBed/Classes/OldTests/FormatsTest.cpp
	Projects/TestBed/Classes/OldTests/FormatsTest.h
	Projects/TestBed/Classes/OldTests/KeyedArchiveTest.h
	Projects/TestBed/Classes/OldTests/MaterialCompilerTest.cpp
	Projects/TestBed/Classes/OldTests/MaterialCompilerTest.h
	Projects/TestBed/Classes/OldTests/OpenGLES30FormatTest.cpp
	Projects/TestBed/Classes/OldTests/OpenGLES30FormatTest.h
	Projects/TestBed/Classes/OldTests/PVRTest.cpp
	Projects/TestBed/Classes/OldTests/PVRTest.h
	Projects/TestBed/Classes/OldTests/ParseTextTest.cpp
	Projects/TestBed/Classes/OldTests/ParseTextTest.h
	Projects/TestBed/Classes/OldTests/RectSpriteTest.cpp
	Projects/TestBed/Classes/OldTests/RectSpriteTest.h
	Projects/TestBed/Classes/OldTests/SampleTest.cpp
	Projects/TestBed/Classes/OldTests/SampleTest.h
	Projects/TestBed/Classes/OldTests/SceneSystemTest.cpp
	Projects/TestBed/Classes/OldTests/SceneSystemTest.h
	Projects/TestBed/Classes/OldTests/SoundTest.cpp
	Projects/TestBed/Classes/OldTests/SoundTest.h
	Projects/TestBed/Classes/OldTests/ThreadSyncTest.cpp
	Projects/TestBed/Classes/OldTests/ThreadSyncTest.h
	Projects/TestBed/Classes/OldTests/UIMovieTest.cpp
	Projects/TestBed/Classes/OldTests/UIMovieTest.h
	Projects/TestBed/Classes/OldTests/UIParticlesTest.cpp
	Projects/TestBed/Classes/OldTests/UIParticlesTest.h
	Projects/TestBed/Classes/Tests/ComponentsTest.cpp
	Projects/TestBed/Classes/Tests/ComponentsTest.h
	Projects/TestBed/Classes/Tests/Cpp14.cpp
	Projects/TestBed/Classes/Tests/Cpp14.h
	Projects/TestBed/Classes/Tests/FunctionBindSingalTest.cpp
	Projects/TestBed/Classes/Tests/FunctionBindSingalTest.h
	Projects/TestBed/Classes/Tests/HashMapTest.cpp
	Projects/TestBed/Classes/Tests/HashMapTest.h
	Projects/TestBed/Classes/Tests/JNITest.cpp
	Projects/TestBed/Classes/Tests/JNITest.h
	Projects/TestBed/Classes/Tests/JobManagerTest.cpp
	Projects/TestBed/Classes/Tests/JobManagerTest.h
	Projects/TestBed/Classes/Tests/KeyedArchiveYamlTest.h
	Projects/TestBed/Classes/Tests/MathTest.cpp
	Projects/TestBed/Classes/Tests/MathTest.h
	Projects/TestBed/Classes/Tests/MemoryAllocatorsTest.cpp
	Projects/TestBed/Classes/Tests/MemoryAllocatorsTest.h
	Projects/TestBed/Classes/Tests/SaveImageTest.cpp
	Projects/TestBed/Classes/Tests/SaveImageTest.h
	Projects/TestBed/Classes/Tests/SplitTest.cpp
	Projects/TestBed/Classes/Tests/SplitTest.h
	Projects/TestBed/Classes/Tests/StringFormatTest.cpp
	Projects/TestBed/Classes/Tests/StringFormatTest.h
	Projects/TestBed/Classes/Tests/TextSizeTest.cpp

Former-commit-id: a4ba8980376f8237180d31464998f641b07fc921

+++ Sources/Internal/DLC/Patcher/PatchFile.cpp
@@ -26,6 +26,7 @@
     SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 =====================================================================================*/
 
+
 #include "PatchFile.h"
 #include "FileSystem/File.h"
 #include "FileSystem/FileList.h"
@@ -116,7 +117,7 @@ bool PatchInfo::ReadString(File* file, String &str)
 {
     bool ret = false;
     uint32 len = 0;
-    uint32 rlen = file->Read(&len);
+    size_t rlen = file->Read(&len);
 
     if(rlen == sizeof(len))
     {
@@ -124,7 +125,7 @@ bool PatchInfo::ReadString(File* file, String &str)
         {
             String tmpStr;
             tmpStr.resize(len);
-            rlen = file->Read(&tmpStr[0], tmpStr.size());
+            rlen = file->Read(&tmpStr[0], static_cast<uint32>(tmpStr.size()));
 
             if(rlen == tmpStr.size())
             {
